requirements.txt
Django>=5.1,<5.2
psycopg2-binary>=2.9
django-ckeditor-5>=0.2.13
Pillow>=10.0
python-dotenv>=1.0
whitenoise>=6.5
django-compressor>=4.4
gunicorn>=22.0
